ocicrypt: add feature "block-cipher"
Add feature "block-cipher" to make image layer encryption/decryption.
