+++ Cargo.toml
@@ -10,19 +10,19 @@ edition = "2018"
 
 [dependencies]
 anyhow = ">=1.0"
-aes = ">=0.8"
+aes = { version = ">=0.8", optional = true }
 base64 = "0.13"
-base64-serde = "0.6"
-ctr = ">=0.9"
-hmac = ">=0.12"
+base64-serde = { version = "0.6", optional = true }
+ctr = { version = ">=0.9", optional = true }
+hmac = { version = ">=0.12", optional = true }
 josekit = { version = ">=0.7", optional = true }
 lazy_static = ">=1.4"
-openssl = { version = ">=0.10", features = ["vendored"] }
-pin-project-lite = "0.2.9"
+openssl = { version = ">=0.10", features = ["vendored"], optional = true }
+pin-project-lite = { version = "0.2.9", optional = true }
 prost = { version = ">=0.11.0", optional = true }
 serde = { version = ">=1.0", features = ["derive"] }
 serde_json = ">=1.0"
-sha2 = ">=0.10"
+sha2 = { version = ">=0.10", optional = true }
 tokio = { version = "1.17.0", features = ["rt-multi-thread"], optional = true }
 tonic = { version = ">=0.8.0", optional = true }
 attestation_agent = { git = "https://github.com/confidential-containers/attestation-agent", rev = "b45b0f8", optional = true  }
@@ -34,9 +34,10 @@ tonic-build = {version = "0.8.0", optional = true }
 aes-gcm = { version = "0.10" }
 
 [features]
-default = ["keywrap-jwe", "keywrap-keyprovider-cmd"]
+default = ["block-cipher", "keywrap-jwe", "keywrap-keyprovider-cmd"]
 eaa_kbc = ["keywrap-keyprovider-native", "attestation_agent/eaa_kbc"]
 async-io = ["tokio"]
+block-cipher = ["aes", "base64-serde", "ctr", "hmac", "openssl", "pin-project-lite", "sha2"]
 keywrap-jwe = ["josekit"]
 keywrap-keyprovider = []
 keywrap-keyprovider-cmd = ["keywrap-keyprovider"]
